feat(calculator) : 입력된 숫자들의 합
