+++ src/test/java/step1/CalculatorTest.java
@@ -5,12 +5,19 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class CalculatorTest {
+    @Test
+    void SumTest() {
+        Calculator calculator = new Calculator(Arrays.asList(1, 2, 3));
+        assertThat(calculator.getSum()).isEqualTo(6);
+
+    }
 
     @ParameterizedTest
     @ValueSource(strings = {
@@ -20,7 +27,7 @@ class CalculatorTest {
             "3:4:5",
             "3,4,5"
     })
-    void splitTest(String input) {
+    void SplitTest(String input) {
         List<String> inputWords = StringUtil.splitStrings(input);
         assertThat(inputWords).containsExactly("3", "4", "5");
     }
@@ -28,7 +35,7 @@ class CalculatorTest {
     @Test
     void StringToIntTest() {
         List<String> inputWords = Arrays.asList("3", "4", "5");
-        List<Integer> parsedNumber = StringUtil.parsePositiveInt(inputWords);
+        List<Integer> parsedNumber = NumberUtil.parsePositiveInt(inputWords);
         assertThat(parsedNumber).containsExactly(3, 4, 5);
     }
 
@@ -37,7 +44,7 @@ class CalculatorTest {
             "e", "-3", ":"
     })
     void StringToIntThrowExceptionTest(String input) {
-        List<String> inputWords = Arrays.asList(input);
-        assertThatThrownBy(() -> StringUtil.parsePositiveInt(inputWords)).isInstanceOf(RuntimeException.class);
+        List<String> inputWords = Collections.singletonList(input);
+        assertThatThrownBy(() -> NumberUtil.parsePositiveInt(inputWords)).isInstanceOf(RuntimeException.class);
     }
 }
